backend/requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==6.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.2
deep-translator==1.11.4
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
evaluate==0.4.3
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.12
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imbalanced-learn==0.13.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
langdetect==1.0.9
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
openai==1.66.3
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.0
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pyparsing==3.2.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentiment-analysis==0.0.1
setuptools==76.0.0
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
textblob==0.19.0
threadpoolctl==3.6.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
